Major update. Global player randomization; major code cleanup, organization, and documentation.
Changed variable names to be more consistent:
-Humans are human players
-AIs are computer players
-Both are considered Players
-Variables are now named accordingly

The game # each human player is in is now randomized (using the Fisher-yates shuffle, credit: https://bost.ocks.org/mike/shuffle/
-Previously the first n humans were in game 1, then human (n+1) to human (2n) were in game 2, etc

A lot of comments were added and cleaned up throughout the code for readability.
-Some information that was already there is now presented in a more unified format.

HWerror function made slightly more intuitive

The results printing code is greatly improved. It used to manually print different strings with different amount of spaces for padding based on how high the player ID was. Now, it uses python's str.rjust() function to dynamically resize the strings based on the player count, evenly spacing the output for up to an infinite amount of players.