Added ConfigAlternateHumanTeamAssignmentAlgorithm option
This option allows the user to choose to use a different algorithm for assigning humans to teams. It chooses a random open slot on teach team with an equal chance for each open slot to be selected, instead of just giving everyone a 50/50 chance of team A or team B.

The meaning of this is that the alternate algorithm has a slight bias towards more even teams in terms of number of humans. For example, in a 3v3 with 6 humans, when the first player is put on team 1, the second player has a 2/5 chance of being put on team 1 and a 3/5 chance of being put on team 2.

This new algorithm is still somewhat WIP and should still be stress tested for further bugs.