Added ConfigForceHumanAlliance option
This option forces all human players to be on the same team in every game if set to 1. If ConfigMaxHumansPerGame is higher than 3 when this is set, it is set to 3.

Also, you can set ConfigForceHumanAlliance to -1 to make the algorithm always put humans on randomized balanced teams against each other, where team 1 always gets the extra player if applicable.